Add Accelerate Data Structure

+++ SandBox/src/EditorLayer/RayTracingEditorLayer.cpp
@@ -11,8 +11,8 @@ RayTracingEditorLayer::RayTracingEditorLayer() :
 {
 	CreateMultiSphereScene();
 	//CreatFourSphereScene();
-	m_spRayTracingCamera->SetPosition(glm::vec3(-10, 6.5, 7));
-	m_spRayTracingCamera->SetForwardDirection(glm::vec3(0.7, -0.6, -0.5));
+	m_spRayTracingCamera->SetPosition(glm::dvec3(-10, 6.5, 7));
+	m_spRayTracingCamera->SetForwardDirection(glm::dvec3(0.7, -0.6, -0.5));
 }
 
 void RayTracingEditorLayer::OnAttach()
@@ -130,10 +130,10 @@ void RayTracingEditorLayer::OnImGuiRender()
 	//auto& mapObjcet = m_spObjectContainer->GetAllObject();
 	//auto& mapMaterial = m_spRayTracingScene->GetMaterials();
 
-	//glm::vec3 position(0.f);
+	//glm::dvec3 position(0.f);
 	//float radius(0.f);
 
-	//glm::vec3 albedo(0.f);
+	//glm::dvec3 albedo(0.f);
 	//float roughness(0.f);
 	//float metallic(0.f);
 	//for (auto& object : mapObjcet)
@@ -197,11 +197,11 @@ void RayTracingEditorLayer::OnImGuiRender()
 void RayTracingEditorLayer::CreatFourSphereScene()
 {
 	auto spMaterial0 = CreateRef<Lambertian>(0);
-	spMaterial0->SetAlbedo(glm::vec3(0.8f, 0.8f, 0.f));
+	spMaterial0->SetAlbedo(glm::dvec3(0.8f, 0.8f, 0.f));
 	m_spRayTracingScene->AddMaterial(spMaterial0);
 
 	auto spMaterial1 = CreateRef<Lambertian>(1);
-	spMaterial1->SetAlbedo(glm::vec3(0.1, 0.2, 0.5));
+	spMaterial1->SetAlbedo(glm::dvec3(0.1, 0.2, 0.5));
 	m_spRayTracingScene->AddMaterial(spMaterial1);
 
 	auto spMaterial2 = CreateRef<Dielectric>(2);
@@ -209,36 +209,36 @@ void RayTracingEditorLayer::CreatFourSphereScene()
 	m_spRayTracingScene->AddMaterial(spMaterial2);
 
 	auto spMaterial3 = CreateRef<Metal>(3);
-	spMaterial3->SetAlbedo(glm::vec3(0.8f, 0.6f, 0.2f));
+	spMaterial3->SetAlbedo(glm::dvec3(0.8f, 0.6f, 0.2f));
 	spMaterial3->SetRoughness(0.f);
 	m_spRayTracingScene->AddMaterial(spMaterial3);
 
 	auto spSpherePrimitive0 = CreateRef<Sphere>(0);
-	spSpherePrimitive0->SetPosition(glm::vec3(0.f, -100.5f, -1.f));
+	spSpherePrimitive0->SetPosition(glm::dvec3(0.f, -100.5f, -1.f));
 	spSpherePrimitive0->SetRadius(100.0f);
 	spSpherePrimitive0->SetMaterialID(0);
 	m_spObjectContainer->AddObject(spSpherePrimitive0);
 
 	auto spSpherePrimitive1 = CreateRef<Sphere>(1);
-	spSpherePrimitive1->SetPosition(glm::vec3(0.f, 0.f, -1.f));
+	spSpherePrimitive1->SetPosition(glm::dvec3(0.f, 0.f, -1.f));
 	spSpherePrimitive1->SetRadius(0.5f);
 	spSpherePrimitive1->SetMaterialID(1);
 	m_spObjectContainer->AddObject(spSpherePrimitive1);
 
 	auto spSpherePrimitive2 = CreateRef<Sphere>(2);
-	spSpherePrimitive2->SetPosition(glm::vec3(-1.f, 0.f, -1.f));
+	spSpherePrimitive2->SetPosition(glm::dvec3(-1.f, 0.f, -1.f));
 	spSpherePrimitive2->SetRadius(0.5f);
 	spSpherePrimitive2->SetMaterialID(2);
 	m_spObjectContainer->AddObject(spSpherePrimitive2);
 
 	auto spSpherePrimitive3 = CreateRef<Sphere>(3);
-	spSpherePrimitive3->SetPosition(glm::vec3(-1.f, 0.f, -1.f));
+	spSpherePrimitive3->SetPosition(glm::dvec3(-1.f, 0.f, -1.f));
 	spSpherePrimitive3->SetRadius(-0.4f);
 	spSpherePrimitive3->SetMaterialID(2);
 	m_spObjectContainer->AddObject(spSpherePrimitive3);
 
 	auto spSpherePrimitive4 = CreateRef<Sphere>(4);
-	spSpherePrimitive4->SetPosition(glm::vec3(1.f, 0.f, -1.f));
+	spSpherePrimitive4->SetPosition(glm::dvec3(1.f, 0.f, -1.f));
 	spSpherePrimitive4->SetRadius(0.5f);
 	spSpherePrimitive4->SetMaterialID(3);
 	m_spObjectContainer->AddObject(spSpherePrimitive4);
@@ -255,15 +255,15 @@ void RayTracingEditorLayer::CreateMultiSphereScene()
 		for (int b = -5; b < 5; b++)
 		{
 			float fChooseMat = Random::Float();
-			glm::vec3 center(a + 0.9 * Random::Float(), 0.2, b + 0.9 * Random::Float());
+			glm::dvec3 center(a + 0.9 * Random::Float(), 0.2, b + 0.9 * Random::Float());
 
-			if (glm::length(center - glm::vec3(4, 0.2, 0)) > 0.9)
+			if (glm::length(center - glm::dvec3(4, 0.2, 0)) > 0.9)
 			{
 
 				if (fChooseMat < 0.8)
 				{
 					//diffuse
-					auto moveCenter = center + glm::vec3(0.f, 0.5f * Random::Float(), 0.f);
+					auto moveCenter = center + glm::dvec3(0.f, 0.5f * Random::Float(), 0.f);
 					auto spSphere = CreateRef<MovingSphere>(iObjectNum++);
 					spSphere->SetPosition(center);
 					spSphere->SetMovePosition(moveCenter);
@@ -287,7 +287,7 @@ void RayTracingEditorLayer::CreateMultiSphereScene()
 
 					//metal
 					auto spLambertian = CreateRef<Metal>(iMaterialNum++);
-					spLambertian->SetAlbedo(0.5f * Random::Vec3());
+					spLambertian->SetAlbedo(0.5 * Random::Vec3());
 					spLambertian->SetRoughness(Random::Float() * 0.5);
 					m_spRayTracingScene->AddMaterial(spLambertian);
 				}
@@ -309,16 +309,16 @@ void RayTracingEditorLayer::CreateMultiSphereScene()
 	}
 
 	auto spGroundSphere = CreateRef<Sphere>(iObjectNum++);
-	spGroundSphere->SetPosition(glm::vec3(0, -1000, 0));
+	spGroundSphere->SetPosition(glm::dvec3(0, -1000, 0));
 	spGroundSphere->SetRadius(1000);
 	spGroundSphere->SetMaterialID(iMaterialNum);
 	m_spObjectContainer->AddObject(spGroundSphere);
 	auto spGroundMaterial = CreateRef<Lambertian>(iMaterialNum++);
-	spGroundMaterial->SetAlbedo(glm::vec3(0.5f));
+	spGroundMaterial->SetAlbedo(glm::dvec3(0.5f));
 	m_spRayTracingScene->AddMaterial(spGroundMaterial);
 
 	auto spDielectricSphere = CreateRef<Sphere>(iObjectNum++);
-	spDielectricSphere->SetPosition(glm::vec3(-4, 1, 0));
+	spDielectricSphere->SetPosition(glm::dvec3(-4, 1, 0));
 	spDielectricSphere->SetRadius(1);
 	spDielectricSphere->SetMaterialID(iMaterialNum);
 	m_spObjectContainer->AddObject(spDielectricSphere);
@@ -327,21 +327,21 @@ void RayTracingEditorLayer::CreateMultiSphereScene()
 	m_spRayTracingScene->AddMaterial(spDielectric);
 
 	auto spLambertianSphere = CreateRef<Sphere>(iObjectNum++);
-	spLambertianSphere->SetPosition(glm::vec3(0, 1, 0));
+	spLambertianSphere->SetPosition(glm::dvec3(0, 1, 0));
 	spLambertianSphere->SetRadius(1);
 	spLambertianSphere->SetMaterialID(iMaterialNum);
 	m_spObjectContainer->AddObject(spLambertianSphere);
 	auto spLambertian = CreateRef<Lambertian>(iMaterialNum++);
-	spLambertian->SetAlbedo(glm::vec3(0.4, 0.2, 0.1));
+	spLambertian->SetAlbedo(glm::dvec3(0.4, 0.2, 0.1));
 	m_spRayTracingScene->AddMaterial(spLambertian);
 
 	auto spMetalSphere = CreateRef<Sphere>(iObjectNum++);
-	spMetalSphere->SetPosition(glm::vec3(4, 1, 0));
+	spMetalSphere->SetPosition(glm::dvec3(4, 1, 0));
 	spMetalSphere->SetRadius(1);
 	spMetalSphere->SetMaterialID(iMaterialNum);
 	m_spObjectContainer->AddObject(spMetalSphere);
 	auto spMetal = CreateRef<Metal>(iMaterialNum++);
-	spMetal->SetAlbedo(glm::vec3(0.7, 0.6, 0.5));
+	spMetal->SetAlbedo(glm::dvec3(0.7, 0.6, 0.5));
 	m_spRayTracingScene->AddMaterial(spMetal);
