+++ SandTable/src/SandTable/RayTracing/Material/Lambertian.cpp
@@ -11,23 +11,23 @@ Lambertian::Lambertian(uint32_t uiMaterialID) :
 {
 }
 
-bool Lambertian::Scatter(const Ray& rayIn, const HitRecord& hitRecord, glm::vec3& attenuation, Ray& rayOut) const
+bool Lambertian::Scatter(const Ray& rayIn, const HitRecord& hitRecord, glm::dvec3& attenuation, Ray& rayOut) const
 {
 	attenuation = m_vec3Albedo;
 	rayOut.Step = rayIn.Step;
-	rayOut.Origin = glm::dot(rayOut.Direction, hitRecord.WorldNormal) < 0 ? hitRecord.WorldPosition - hitRecord.WorldNormal * 0.00001f :
-		hitRecord.WorldPosition + hitRecord.WorldNormal * 0.00001f;
+	rayOut.Origin = glm::dot(rayOut.Direction, hitRecord.WorldNormal) < 0 ? hitRecord.WorldPosition - hitRecord.WorldNormal * 0.00001 :
+		hitRecord.WorldPosition + hitRecord.WorldNormal * 0.00001;
 	rayOut.Direction = hitRecord.WorldNormal + Random::UnitSphere();
 	return true;
 }
 
-void Lambertian::SetAlbedo(const glm::vec3& vec3Albedo)
+void Lambertian::SetAlbedo(const glm::dvec3& vec3Albedo)
 {
 	m_bDirty = true;
 	m_vec3Albedo = vec3Albedo;
 }
 
-const glm::vec3& Lambertian::GetAlbedo() const
+const glm::dvec3& Lambertian::GetAlbedo() const
 {
 	return m_vec3Albedo;
 }
